Initial workload to make cross platform.
A lot of this is just #define hacks and a couple of typedefs.

I removed the helper functions to check if a CPU is compable with various things, as it was a high workload to work out if it was compatible, but there was no use of it.

Time to go back to testing windows compilation